// src/app/model/Peofile.ts
export class Profile {
    Details: string;
    Gender: string;
    ImageID:  number;
    Language:  string;
    LastOnlineTime: Date;
    Location:  string;
    Name: string;
    Nationality:  string;
    SpecialisationIn:  string;
    StudiesAt:  string;
    UserID:  string;
    UserLoginStatus:  string;
    UserName:  string;
    UserTimeZone:  string;
    UserType:  string;
    Duration:  string;
    LocalOffset:  number;
    LocalTime:  Date;
    image: { darkImage: string; lightImage: string ;darkImageErr: any; lightImageErr: string};
}